+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validation for `defaultApproveNotificationEmailElementId`, `defaultClarificationNotificationEmailElementId` and `defaultDenyNotificationEmailElementId` properties
+
 ## [0.19.11] - 2022-08-28
 
 ### Changed